Scotty request size checks.
Thanks to @fpringle, we discovered that the max. size was not taken
into account if it was less than the chunk size. We'd receive the
first chunk and only then check the size, which wouldn't detect a
larger request until the first chunk was processed.

A request limit should be taken into account even for chunks larger
than the limit.

+ Some cleanup to make the function more readable.

+ test cases: should also pass for 1024, not just 1023.